ai_engine/env_requirements.txt
paho-mqtt
pillow
numpy
tensorflow